add RP2040 port with rp2040js simulation harness
- platform/rp2040/: CMakeLists, main.cpp (UART protocol), and rp2040js
  Node test harness that boots the firmware and captures audio to WAV
- Fix int32_t/int literal mismatches for arm-none-eabi (int32_t = long):
  std::max/min calls in AudioProcessor and KlattSynthesizer, and clamp11
  call sites in SpeechRenderer and TtsEngine (explicit <int32_t> template
  arg avoids ambiguity on both ARM and x86)
- Replace SymbolsTable unordered_map with a flash-resident const array +
  FindSymbol() linear search; saves ~4KB heap and moves all symbol data
  to XIP flash; update all call sites including DumpStages test
- Add #include <unordered_map> directly to Phonemizer (was transitive)
- PICO_CXX_ENABLE_EXCEPTIONS=1 before pico_sdk_init(); scope -fno-rtti
  to CXX only; set PICO_STACK_SIZE=0x1000 to fit SCRATCH_Y (4KB)

+++ src/AudioProcessor.cpp
@@ -1259,7 +1259,7 @@ namespace SharpVox {
                     if (curPhon != _SIL_ && d < 8 / kFrameTime) {
                         d = 8 / kFrameTime;
                     }
-                    _durBuf[i] = (int16_t)std::max(d, 1);
+                    _durBuf[i] = (int16_t)std::max(d, (int32_t)1);
                     goto DURATION_DONE_END;
                 }
             }
@@ -1495,7 +1495,7 @@ namespace SharpVox {
                 if (curPhon != _SIL_ && d < 8 / kFrameTime) {
                     d = 1;
                 }
-                _durBuf[i] = (int16_t)std::max(d, 1);
+                _durBuf[i] = (int16_t)std::max(d, (int32_t)1);
             }
 
         DURATION_DONE_END:;
@@ -1956,7 +1956,7 @@ namespace SharpVox {
         }
 
         for (int32_t i = start; i <= end; i++) {
-            _durBuf[i] = (int16_t)std::max(1, (_durBuf[i] * pct + 50) / 100);
+            _durBuf[i] = (int16_t)std::max((int32_t)1, (_durBuf[i] * pct + 50) / 100);
         }
     }
